backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
gunicorn==20.1.0
dj-database-url==1.0.0
whitenoise==6.4.0
django-environ==0.10.0
Pillow==9.4.0
python-dotenv>=0.19.2
django-environ==0.10.0
django-filter==23.2
